Move builtin methods to stdlib module
This is just moving code around, no functional changes.
