+++ src/stdlib.rs
@@ -12,7 +12,7 @@ use std::rc::Rc;
 
 use crate::error::{IntoError, Result};
 use crate::eval::Evaluator;
-use crate::runtime::{builtin_function, FunctionCall, Value};
+use crate::runtime::{builtin_function, builtin_method, FunctionCall, MethodCall, Value};
 
 builtin_function!(
     "std.read_file_utf8",
@@ -49,4 +49,70 @@ pub fn initialize() -> Rc<Value> {
     );
 
     Rc::new(Value::Dict(builtins))
+}
+
+builtin_method!("Dict.len", const DICT_LEN, builtin_dict_len);
+fn builtin_dict_len(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("Dict.len", &[])?;
+    let dict = call.receiver.expect_dict();
+    Ok(Rc::new(Value::Int(dict.len() as _)))
+}
+
+builtin_method!("List.len", const LIST_LEN, builtin_list_len);
+fn builtin_list_len(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("List.len", &[])?;
+    let list = call.receiver.expect_list();
+    Ok(Rc::new(Value::Int(list.len() as _)))
+}
+
+builtin_method!("Set.len", const SET_LEN, builtin_set_len);
+fn builtin_set_len(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("Set.len", &[])?;
+    let set = call.receiver.expect_set();
+    Ok(Rc::new(Value::Int(set.len() as _)))
+}
+
+builtin_method!("String.len", const STRING_LEN, builtin_string_len);
+fn builtin_string_len(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("String.len", &[])?;
+    let string = call.receiver.expect_string();
+    Ok(Rc::new(Value::Int(string.len() as _)))
+}
+
+builtin_method!("Dict.contains", const DICT_CONTAINS, builtin_dict_contains);
+fn builtin_dict_contains(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("Dict.contains", &["key"])?;
+    let dict = call.receiver.expect_dict();
+    let needle = &call.call.args[0].1;
+    Ok(Rc::new(Value::Bool(dict.contains_key(needle))))
+}
+
+builtin_method!("List.contains", const LIST_CONTAINS, builtin_list_contains);
+fn builtin_list_contains(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call
+        .check_arity_static("List.contains", &["element"])?;
+    let list = call.receiver.expect_list();
+    let needle = &call.call.args[0].1;
+    Ok(Rc::new(Value::Bool(list.contains(needle))))
+}
+
+builtin_method!("Set.contains", const SET_CONTAINS, builtin_set_contains);
+fn builtin_set_contains(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call.check_arity_static("Set.contains", &["element"])?;
+    let set = call.receiver.expect_set();
+    let needle = &call.call.args[0].1;
+    Ok(Rc::new(Value::Bool(set.contains(needle))))
+}
+
+builtin_method!("Dict.get", const DICT_GET, builtin_dict_get);
+fn builtin_dict_get(_eval: &mut Evaluator, call: MethodCall) -> Result<Rc<Value>> {
+    call.call
+        .check_arity_static("Dict.get", &["key", "default"])?;
+    let dict = call.receiver.expect_dict();
+    let key = &call.call.args[0].1;
+    let default = &call.call.args[1].1;
+    match dict.get(key) {
+        Some(v) => Ok(v.clone()),
+        None => Ok(default.clone()),
+    }
 }
